+++ src/managers/UnlockPanelManager/UnlockPanelManager.ts
@@ -15,6 +15,7 @@ import {
   UnlockPanelManagerInitParamsType
 } from './UnlockPanelManager.types';
 import { SidePanelBaseManager } from '../internal/SidePanelBaseManager';
+import { getIsSimpleLoginCallback } from './helpers/getIsSimpleLoginCallback';
 import { getProvidersList } from './helpers/getProviderList';
 
 interface IUnlockPanelManagerData {
@@ -112,7 +113,7 @@ export class UnlockPanelManager extends SidePanelBaseManager<
     }
 
     try {
-      if (this.isSimpleLoginCallback(UnlockPanelManager.loginHandler)) {
+      if (getIsSimpleLoginCallback(UnlockPanelManager.loginHandler)) {
         const provider = await ProviderFactory.create({ type, anchor });
         await provider.login();
         UnlockPanelManager.loginHandler();
@@ -131,12 +132,5 @@ export class UnlockPanelManager extends SidePanelBaseManager<
 
   private readonly handleCancelLogin = async () => {
     await ProviderFactory.destroy();
-  };
-
-  private readonly isSimpleLoginCallback = (
-    login: LoginHandlerType
-  ): login is () => void => {
-    const takesZeroArguments = login.length === 0;
-    return takesZeroArguments;
   };
 }